Refactored the Service code
